added one to many relationship in the database

+++ app/routers/post.py
@@ -1,4 +1,4 @@
-from .. import schemas, utils, models, oauth2
+from .. import schemas, models, oauth2
 from fastapi import FastAPI, Response, status, HTTPException, Depends, APIRouter
 from sqlalchemy.orm import Session
 from ..database import get_db
@@ -11,22 +11,24 @@ router = APIRouter(
 
 @router.get("/", response_model=List[schemas.PostResponse])
 def get_posts(db: Session=Depends(get_db), 
-              current_user: int = Depends(oauth2.get_current_user)):
-    posts = db.query(models.Post).all()
+              current_user: dict = Depends(oauth2.get_current_user)):
+    
+    posts = db.query(models.Post).filter(models.Post.id == current_user.id).all()
     return posts
 
 
 
 @router.post("/", status_code=status.HTTP_201_CREATED, response_model=schemas.PostResponse)
 def create_posts(post: schemas.PostCreate, db: Session=Depends(get_db), 
-                 current_user: int = Depends(oauth2.get_current_user)):
+                 current_user: dict = Depends(oauth2.get_current_user)):
     """ 
     The user needs to be logged in to his account to have access to post.
     So we create a dependency with the oauth2 get_current_user which will raise credentials exception 
     if the user is not logged in or the time of the token has expired 
+    We also have to include the owner_id into to the post before adding it to the user 
     """
     # '**' unpacks that dictionary into arguments.
-    new_post = models.Post(**post.model_dump())
+    new_post = models.Post(owner_id = current_user.id, **post.model_dump())
     db.add(new_post)
     db.commit()
     db.refresh(new_post)
@@ -36,7 +38,7 @@ def create_posts(post: schemas.PostCreate, db: Session=Depends(get_db),
 
 @router.get("/{id}", response_model=schemas.PostResponse)
 def get_posts(id: int, db: Session=Depends(get_db),
-              current_user: int = Depends(oauth2.get_current_user)):
+              current_user: dict = Depends(oauth2.get_current_user)):
     
     post = db.query(models.Post).filter(models.Post.id == id).first()
     if not post: 
@@ -47,7 +49,7 @@ def get_posts(id: int, db: Session=Depends(get_db),
 
 @router.delete("/{id}", status_code=status.HTTP_204_NO_CONTENT)
 def delete_post(id: int, db: Session=Depends(get_db),
-                current_user: int = Depends(oauth2.get_current_user)):
+                current_user: dict = Depends(oauth2.get_current_user)):
     
     post_query = db.query(models.Post).filter(models.Post.id == id)
 
@@ -55,6 +57,10 @@ def delete_post(id: int, db: Session=Depends(get_db),
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, 
                             detail=f"The post with id: {id} not found") 
     else:
+        if post_query.first().owner_id != current_user.id:
+            raise HTTPException(status_code=status.HTTP_403_FORBIDDEN,
+                                detail="Not Authorized to perform requested action")
+
         post_query.delete(synchronize_session=False)
         db.commit()
     return Response(status_code=status.HTTP_204_NO_CONTENT)
@@ -63,7 +69,11 @@ def delete_post(id: int, db: Session=Depends(get_db),
 
 @router.put("/{id}", response_model=schemas.PostResponse)
 def update_posts(id: int, post: schemas.PostBase, db: Session=Depends(get_db),
-                 current_user: int = Depends(oauth2.get_current_user)):
+                 current_user: dict = Depends(oauth2.get_current_user)):
+    """
+    Only the user who is logged in can have access to updating his post 
+    We get this from the dependency to get the current_user which verifies that the token is valid or not.
+    """
 
     post_query = db.query(models.Post).filter(models.Post.id == id)
 
@@ -71,6 +81,10 @@ def update_posts(id: int, post: schemas.PostBase, db: Session=Depends(get_db),
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, 
                             detail=f"The post with id: {id} not found")
     else:
+        if post_query.first().owner_id != current_user.id:
+            raise HTTPException(status_code=status.HTTP_403_FORBIDDEN,
+                                detail="Not Authorized to perform requested action")
+        
         post_query.update(post.model_dump(), synchronize_session=False)
         db.commit()
         modified_post = post_query.first()
